windows: set default editor to notepad
The previous default - "edit" - doesn't exist on windows 7 and later,
and there's no other console editor by default, while notepad exists
since at least XP or possibly earlier.

This may be an issue if the user ssh into the machine without GUI,
but hopefully notepad will fail to launch in such case, or worst case,
the user will learn to set EDITOR for the next time...

Also, on windows, remove the +<LINE> argument, as notepad doesn't
support it.

The default can be overridden normally using VISUAL/EDITOR/LESSEDIT.

Unrelated to windows, with LESSOPEN it displays a warning which needs
to be dismissed manually ("press RETURN"), and this feels a bit meh,
especially when there's no way to cancel if the user thinks the
warning has a point, i.e. the editor will be launched anyway.