windows: SGR processor: support OSC sequences
Previously the windows SGR processor detected ESC[ as CSI, and other
ESCx were ignored. This resulted in OSC sequences ESC]...ST to be
printed literally after the ESC].

Now OSC sequences are detected and passthrough to the terminal if VT
is enabled, else ignored till the end of sequence.

This makes OSC-8 work on Windows 10+ regardless of sgr_mode (-Da),
and display as plain text where VT is not supported - like in any
terminal which doesn't support OSC-8 but still ignores unknown OSC.