windows: utf8: don't depend on console output CP
On Windows, with UTF-8 output, if the console output CP (COCP) is not
UTF-8 (65001), then we convert output stings to wchar_t and print them
as wide strings, else we write the output bytes and expect the console
to digest it correctly.

This generally works, except for two cases:

- If we're piping from another app, then that app may change the COCP
  when it starts and when it exits, but we only test it on startup,
  which can lead to incorrect knowledge at runtime and e.g. missing
  conversion (if it was UTF-8 on startup but later something else).

- If we output a string with partial UTF-8 codepoint, then it can't
  be converted, and the output is incorrect. This can happen when
  writing a partial string from a buffer and that buffer is not big
  enough to hold the whole string.

This commit adds a function writeCon_utf8 which takes char* input
and writes it as if the COCP is UTF-8 but regardless of the actual
COCP, and which handles partial codepoints by bufferring them.

With this change, there are now only two values for utf_mode:
0: we don't write UTF-8.
1: we do write UTF-8.

Where previously on Windows it could also be 2 which meant that we
write UTF-8 but the COCP is not UTF-8 (and conversion is needed).

Additionally, the buffer size in WIN32textout is now decoupled from
obuf size at screen.c, and 4096 is chosen for performance, but it
could be as small as 2 (minimum to hold a surrogates pair).